Move main pluginsserver code under src/main
