+++ benchmarks/clickhouse/e2e.timeout.benchmark.ts
@@ -1,6 +1,6 @@
 import { performance } from 'perf_hooks'
 
-import { startPluginsServer } from '../../src/server'
+import { startPluginsServer } from '../../src/main/pluginsServer'
 import { KAFKA_EVENTS_PLUGIN_INGESTION } from '../../src/shared/ingestion/topics'
 import { delay, UUIDT } from '../../src/shared/utils'
 import { ClickHouseEvent, LogLevel, PluginsServerConfig, Queue } from '../../src/types'
