+++ src/index.ts
@@ -1,7 +1,7 @@
 import * as yargs from 'yargs'
 
 import { initApp } from './init'
-import { startPluginsServer } from './server'
+import { startPluginsServer } from './main/pluginsServer'
 import { configHelp, defaultConfig } from './shared/config'
 import { PluginsServerConfig } from './types'
 import { makePiscina } from './worker/piscina'
