+++ tests/postgres/worker.test.ts
@@ -2,8 +2,8 @@ import { PluginEvent } from '@posthog/plugin-scaffold/src/types'
 import IORedis from 'ioredis'
 import { mocked } from 'ts-jest/utils'
 
+import { ServerInstance, startPluginsServer } from '../../src/main/pluginsServer'
 import { loadPluginSchedule } from '../../src/main/services/schedule'
-import { ServerInstance, startPluginsServer } from '../../src/server'
 import Client from '../../src/shared/celery/client'
 import { delay, UUIDT } from '../../src/shared/utils'
 import { LogLevel } from '../../src/types'
