+++ tests/server.test.ts
@@ -1,6 +1,4 @@
-import { PluginEvent } from '@posthog/plugin-scaffold/src/types'
-
-import { startPluginsServer } from '../src/server'
+import { startPluginsServer } from '../src/main/pluginsServer'
 import { LogLevel } from '../src/types'
 import { makePiscina } from '../src/worker/piscina'
 import { resetTestDatabase } from './helpers/sql'
